+++ packages/mason_logger/lib/mason_logger.dart
@@ -3,7 +3,7 @@
 /// Get started at [https://github.com/felangel/mason](https://github.com/felangel/mason) 🧱
 library mason_logger;
 
-export 'src/io.dart';
+export 'src/io.dart' hide ControlCharacter, KeyStroke, readKeyStroke;
 export 'src/level.dart';
 export 'src/link.dart';
 export 'src/mason_logger.dart'
